test: make test suite publicly visible
